Sqlite Overflow pages partially working

+++ src/univix_util.h
@@ -137,8 +137,8 @@ public:
         return ret;
     }
 
-    static int compare(const uint8_t *v1, uint8_t len1, const uint8_t *v2,
-            uint8_t len2, int k = 0) {
+    static int compare(const uint8_t *v1, int len1, const uint8_t *v2,
+            int len2, int k = 0) {
         int lim = (len2 < len1 ? len2 : len1);
         while (k < lim) {
             uint8_t c1 = v1[k];
